Update Ruby test pipeline working while loop
After many iterations the while loop for the pause and running check works.  Had to get the execution date by first getting the first index of json ([0]) and then getting date.

Also had to grep the output for dag state in order to correctly get the value needed for the varaible (it prints more than just the state).  Since working adding a message to show when it exist the while loop.